Created websocket connection with server to implement working compilation and execution of C++ files
Vastly improved the security of the system by incorperating JWT tokens, and also created a websocket connection so that the client is able to send inputs back to the database in order to compile files that require input from the terminal! Also vastly improved the aesthetic of the coding page!